Assignment 1/requirements.txt
pytest
tensorflow
numpy
PyYAML
